+++ NEWS.md
@@ -6,6 +6,15 @@ width: 128px;
 border-radius: 128px;
 " />
 
+## v1.11.5
+
+- Fixes
+    - Fix occasional recreation of comics on docker or network filesystems.
+    - Fix double polling of some libraries.
+    - Fix Codex sometimes hanging on shutdown after Ctrl+C.
+    - Fix PDFs rendering OCR overlay text.
+    - Fix PDFs sometimes not rendering at correct resolution
+
 ## v1.11.4
 
 - Fixes